+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## WIP
+- Added action `Delete`.
+
 ## v1.7.0 (2019-11-26)
 - Added filter `Exif` to filter by image exif data.
 - Placeholder variable properties are now case insensitve.